feat(system): feat(system): DPLAN-0141 Phase 3 — ai_mail audit rescue (email.py split + nesting fix + dead code) + devpulse cleanup (feedback stderr + bypass + README + DPLAN-0141)
